cast

+++ kern/arch/mips/vm/dumbvm.c
@@ -173,8 +173,8 @@ getppages(unsigned long npages)
 				kprintf("here4\n");
 				spinlock_release(&stealmem_lock);
 				kprintf("%d\n", addr);
-				kprintf("%d", addr <= (int)addr_hi);
-				kprintf("%d", addr >= (int)addr_lo);
+				kprintf("%d", (int)addr <= (int)addr_hi);
+				kprintf("%d", (int)addr >= (int)addr_lo);
 				return addr;
 				}
 			}
